pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "corpus-prep"
version = "0.1.0"
description = "Pipeline open-source de preparação de corpus para treinamento de LLMs em PT-BR"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "Heitor Moura" }]
keywords = ["llm", "nlp", "corpus", "training", "dataset", "pt-br", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    # MIME detection
    "magika>=0.5.0",
    # Parsers
    "pymupdf4llm>=0.0.17",
    "docling>=2.0",
    "trafilatura>=1.12",
    # Normalization
    "ftfy>=6.3",
    # Language ID
    "fasttext>=0.9.3",
    # Deduplication
    "text-dedup @ git+https://github.com/ChenghaoMou/text-dedup.git@main",
    "datasketch>=1.6",
    # Output
    "pyarrow>=18.0",
    "duckdb>=1.1",
    # CLI / config / schemas
    "typer>=0.12",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    # Logging / UI
    "structlog>=24.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.13",
    "pre-commit>=4.0",
    "ipykernel>=6.29",
    "jupyter>=1.1",
    # Test fixtures
    "reportlab>=4.0",
    "pillow>=11.0",
]

[project.scripts]
corpus-prep = "corpus_prep.cli:app"

[project.urls]
Homepage = "https://github.com/heitor-am/corpus-prep"
Repository = "https://github.com/heitor-am/corpus-prep"
Issues = "https://github.com/heitor-am/corpus-prep/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/corpus_prep"]

# ---------- Tooling ----------

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle / pyflakes
    "I",             # isort
    "UP",            # pyupgrade
    "B",             # bugbear
    "SIM",           # simplify
    "RUF",           # ruff-specific
    "N",             # pep8-naming
    "PT",            # pytest style
]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_decorators = false
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["pymupdf4llm.*", "docling.*", "trafilatura.*", "fasttext.*", "magika.*", "ftfy.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers --cov=corpus_prep --cov-report=term-missing"
markers = [
    "slow: tests que rodam OCR pesado ou modelos grandes (opt-in via -m slow)",
]

[tool.coverage.run]
source = ["src/corpus_prep"]
branch = true
